Fix format specifier warning introduced couple commits earlier

+++ libtransmission/peer-mgr.c
@@ -2895,9 +2895,9 @@ rechokeDownloads (tr_swarm * s)
         const double mult = MIN (timeSinceCancel, maxHistory) / (double) maxHistory;
         const int inc = maxIncrease * mult;
         maxPeers = s->maxPeers + inc;
-        tordbg (s, "time since last cancel is %"PRId64" -- increasing the "
+        tordbg (s, "time since last cancel is %"PRIdMAX" -- increasing the "
                    "number of peers we're interested in by %d",
-                   (long long)timeSinceCancel, inc);
+                   (intmax_t)timeSinceCancel, inc);
       }
   }
